// web/src/api/generated/models/MemoryDetail.ts
export interface MemoryDetail {
    /**
     * 
     * @type {Array<MemoryAttachmentSummary>}
     * @memberof MemoryDetail
     */
    attachments: Array<MemoryAttachmentSummary>;
    /**
     * 
     * @type {AuthorSummary}
     * @memberof MemoryDetail
     */
    author: AuthorSummary;
    /**
     * 
     * @type {string}
     * @memberof MemoryDetail
     */
    authorId: string;
    /**
     * 
     * @type {string}
     * @memberof MemoryDetail
     */
    body: string;
    /**
     * 
     * @type {ResourceCapabilities}
     * @memberof MemoryDetail
     */
    capabilities: ResourceCapabilities;
    /**
     * 
     * @type {Date}
     * @memberof MemoryDetail
     */
    createdAt: Date;
    /**
     * 
     * @type {Date}
     * @memberof MemoryDetail
     */
    happenedOn: Date | null;
    /**
     * 
     * @type {string}
     * @memberof MemoryDetail
     */
    id: string;
    /**
     * 
     * @type {string}
     * @memberof MemoryDetail
     */
    spaceId: string;
    /**
     * 
     * @type {string}
     * @memberof MemoryDetail
     */
    title: string;
    /**
     * 
     * @type {Date}
     * @memberof MemoryDetail
     */
    updatedAt: Date;
    /**
     * 
     * @type {number}
     * @memberof MemoryDetail
     */
    version: number;
}